refactor(front): move seed lib to lib/dev/dev_seed_user.ts
