+++ front/admin/seed_dev_user.ts
@@ -26,7 +26,7 @@
 
 import * as fs from "fs";
 
-import { seedDevUser, validateSeedConfig } from "@app/lib/seed_dev_user";
+import { seedDevUser, validateSeedConfig } from "@app/lib/dev/dev_seed_user";
 
 async function main() {
   if (process.env.NODE_ENV !== "development") {
